랜딩 페이지 리팩토링 (#8)
* refactor: 페인포인트와 해결 가능 문구 컴포넌트

* refactor: 우리 기능 소개 컴포넌트

* refactor: 나가기/사용해보기 action 버튼 모아보는 컴포넌트

* refactor: 슬로건 컴포넌트

* refactor: 컴포넌트로 페이지 분리

* chore: clean up code

* chore: use client 선언

* chore: console.log 삭제
